+++ packages/rax-waterfall/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-waterfall",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "waterfall component for Rax.",
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
